Cast state variable to long long
In method CalculateVolume(), a cast to one variable is necessary so
the multiplication's type matches the method's return type.
Solve challenge.

+++ hackerrank/box_it.cpp
@@ -36,7 +36,7 @@ class Box {
          * @return volume of the box.
          */
         long long CalculateVolume() {
-            return length * breadth * height;
+            return (long long)length * breadth * height;
         }
 
         /**
